ACM-20404 implement useIsFleetAvailable hook (#4674)
* Implement useIsFleetAvailable hook and test case


* acm plugin sets @stolostron/multicluster-sdk/provider@1 to true


* Delete obsolete code


* Refactor constants for public/internal


* Add documentation


---------
