+++ frontend/packages/multicluster-sdk/src/api/apiRequests.ts
@@ -12,7 +12,7 @@ import {
   Selector,
 } from '@openshift-console/dynamic-plugin-sdk'
 import { selectorToString } from './utils/requirements'
-import { BASE_K8S_API_PATH } from './constants'
+import { BASE_K8S_API_PATH } from '../internal/constants'
 import { getFleetK8sAPIPath } from './useFleetK8sAPIPath'
 
 export type BaseOptions = {
