+++ frontend/packages/multicluster-sdk/src/api/index.ts
@@ -1,5 +1,4 @@
 /* Copyright Contributors to the Open Cluster Management project */
-export * from './FleetSupport'
 export * from './FleetResourceLink'
 export * from './useFleetK8sAPIPath'
 export * from './useFleetK8sWatchResource'
@@ -8,3 +7,4 @@ export * from './search/useMulticlusterSearchWatch'
 export * from './apiRequests'
 export * from './useHubClusterName'
 export * from './useFleetPrometheusPoll'
+export * from './useIsFleetAvailable'
