+++ frontend/packages/multicluster-sdk/src/api/useFleetClusterNames.ts
@@ -1,7 +1,7 @@
 /* Copyright Contributors to the Open Cluster Management project */
 import { K8sResourceCommon, useK8sWatchResource } from '@openshift-console/dynamic-plugin-sdk'
 import { UseFleetClusterNames } from '../types/fleet'
-import { ManagedClusterListGroupVersionKind } from './constants'
+import { ManagedClusterListGroupVersionKind } from '../internal/models'
 
 export const useFleetClusterNames: UseFleetClusterNames = () => {
   const [clusters, loaded, error] = useK8sWatchResource<K8sResourceCommon[]>({
