+++ frontend/packages/multicluster-sdk/src/index.ts
@@ -1,5 +1,4 @@
 /* Copyright Contributors to the Open Cluster Management project */
 export * from './api'
-export * from './extensions'
 export * from './components'
 export * from './types'
