+++ frontend/packages/multicluster-sdk/src/internal/constants.ts
@@ -1,0 +1,8 @@
+/* Copyright Contributors to the Open Cluster Management project */
+
+export const BASE_K8S_API_PATH = '/api/kubernetes'
+export const BASE_FLEET_SEARCH_PATH = '/multicloud/search/resources'
+
+export const MANAGED_CLUSTER_API_PATH = 'managedclusterproxy'
+
+export const LOCAL_CLUSTER_LABEL = 'local-cluster'
