--- frontend/packages/multicluster-sdk/src/internal/hooks/index.ts
@@ -1,2 +1,0 @@
-/* Copyright Contributors to the Open Cluster Management project */
-export * from './useFleetSupport'
