--- frontend/packages/multicluster-sdk/src/internal/hooks/useFleetSupport.ts
@@ -1,7 +1,0 @@
-/* Copyright Contributors to the Open Cluster Management project */
-import { useContext } from 'react'
-import { FleetSupportContext } from '../context/FleetSupportContext'
-
-export function useFleetSupport() {
-  return useContext(FleetSupportContext)
-}
