--- frontend/packages/multicluster-sdk/src/internal/index.ts
@@ -1,3 +1,0 @@
-/* Copyright Contributors to the Open Cluster Management project */
-export * from './hooks'
-export * from './types'
