+++ frontend/packages/multicluster-sdk/src/internal/models.ts
@@ -1,0 +1,20 @@
+/* Copyright Contributors to the Open Cluster Management project */
+import { K8sModel } from '@openshift-console/dynamic-plugin-sdk'
+
+export const ManagedClusterModel: K8sModel = {
+  abbr: 'MC',
+  apiGroup: 'clusterview.open-cluster-management.io',
+  apiVersion: 'v1',
+  crd: true,
+  kind: 'ManagedCluster',
+  label: 'ManagedCluster',
+  labelPlural: 'ManagedClusters',
+  namespaced: false,
+  plural: 'managedclusters',
+}
+
+export const ManagedClusterListGroupVersionKind = {
+  group: 'clusterview.open-cluster-management.io',
+  version: 'v1',
+  kind: 'ManagedCluster',
+}
