--- frontend/packages/multicluster-sdk/src/internal/types/index.ts
@@ -1,2 +1,0 @@
-/* Copyright Contributors to the Open Cluster Management project */
-export * from './multicluster-sdk'
